front + validation + minor bugs fixed

+++ backend/src/DomainScanner.Domain/ValueObjects/HttpResponseObject.cs
@@ -2,7 +2,7 @@
 
 public class HttpResponseObject
 {
-    public int StatusCode { get; set; }
+    public ushort StatusCode { get; set; }
     public bool IsSuccess { get; set; }
     public DateTime CreatedAt { get; set; } = DateTime.UtcNow;
 }
